Changes to Index.html
The changes I made to the index was to add in the images of the credit cards for the users to click on which one they are going to be using, but I sadly have a problem with it that I will have to try to fix tomorrow. Then, I made a label and a input for the user to be able to put in their credit card number in it, and it is required for the user to do so. Lastly I tried to work with putting in a select fieldset for the users to put in their expiration date, but I feel i'm having some trouble with it which I will also have to try to fix tomorrow.